rename Radio component to RadioGroup
