+++ src/script/components/Radio/index.ts
@@ -17,4 +17,4 @@
  *
  */
 
-export * from './Radio';
+export * from './RadioGroup';
